requirements.txt
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
google-api-python-client==2.171.0
playwright==1.52.0
langchain==0.3.25
langchain-openai==0.3.19
python-dotenv==1.1.0
